requirements.txt
aiohttp>=3.9.3
beautifulsoup4>=4.12.3
fake-useragent>=1.5.1
httpx>=0.27.0
numpy>=1.26.4
playwright>=1.44.0
pydantic>=2.7.1
python-dotenv>=1.0.1
PyYAML>=6.0.1
spacy>=3.7.4
typer>=0.12.3
uvicorn>=0.30.1

